Server/studentmanagement/requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==3.2.2
blessed==1.19.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
ci-info==0.3.0
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.5
configobj==5.0.6
configparser==5.3.0
crashtest==0.3.1
cryptography==37.0.4
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-drf-filepond==0.4.1
django-environ==0.9.0
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
django-storages==1.13.1
django-webpack-loader==0.6.0
djangorestframework==3.14.0
dotty-dict==1.3.1
etelemetry==0.3.0
exceptionite==2.2.5
Faker==13.15.1
filelock==3.8.0
fitz==0.0.1.dev2
future==0.18.2
hashids==1.3.1
hfilesize==0.1.0
httplib2==0.20.4
hupper==1.10.3
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.1
Jinja2==3.0.3
jinxed==1.2.0
looseversion==1.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
masonite==4.16.3
masonite-orm==2.18.4
mock==4.0.3
networkx==2.8.6
nibabel==4.0.2
nipype==1.8.4
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pastel==0.2.1
pathlib==1.0.1
pdf2image==1.16.0
pendulum==2.1.2
Pillow==9.2.0
pluggy==1.0.0
prov==2.0.0
py==1.11.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.4.0
pylev==1.4.0
PyMuPDF==1.20.2
pyotp==2.6.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytzdata==2020.1
pyxnat==1.4
rdflib==6.2.0
redis==3.5.3
requests==2.28.1
requests-file==1.5.1
scipy==1.9.1
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tldextract==2.2.3
tomli==2.0.1
traits==6.3.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==5.2.0
